+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -8,6 +8,7 @@
   - `interactiveLegend`: whether legend items are clickable or not. Defaults to true.
   - `legendPosition`: can be `top`, `side` or `bottom`. If not specified, it's calculated based on `mode` and viewport width.
   - `showHeaderControls`: whether the header controls must be visible. Defaults to true.
+- `getColor()` method in chart utils now requires `keys` and `colorScheme` to be passed as separate params.
 
 # 1.3.0